Add files via upload
Written by Brian Lesko for learning purposes
These files are intro exercises to the relevant NYCFlights13 dataset. More info on the data in the README.md for this folder. 

Written in R studio using the markdown format, the HTML is a document preview of the data science and visualization work.