Create README.md
Written by Brian Lesko. This folder contains a data anlysis project in R markdown for learning purposes and is focused around olympics data.

Objectives for this project: 

Learn R data organization and analysis

Answer these questions about the olympic games using data: 
Is there a relationship between a country’s GDP and the number of athletes it sends to an Olympic games?
Is there a host country effect on the number of medals? In other words, do countries earn more medals when they host an Olympic games compared to when they do not host an Olympic games?